add shutdown command
this commit adds a shutdown command. To implement this the CLI was moved into a own thread and the
stopping mechanism is implemented via a semaphore that is locked until the bot should stop. When it
becomes unlocked it will get locked by the main thread which will then safely end all threads.

since `getline` can allocate memory before blocking a cleanup handles was also added for the CLI
thread